+++ R/methods_sampleSelection.R
@@ -3,12 +3,12 @@
 #' @rdname get_coef
 #' @export
 get_coef.selection <- function(model, ...) {
-    # sampleSelection::selection
-    if (as.list(model$call)[[1]] == "selection") {
-        out <- model$estimate
-        # sampleSelection::heckit
-    } else if (as.list(model$call)[[1]] == "heckit") {
+    if ("coefficients" %in% names(model)) {
         out <- model$coefficients
+    } else if ("estimate" %in% names(model)) {
+        out <- model$estimate
+    } else {
+        stop("Model does not have coefficients or estimates.")
     }
     return(out)
 }
@@ -18,11 +18,12 @@ get_coef.selection <- function(model, ...) {
 #' @export
 set_coef.selection <- function(model, coefs, ...) {
     # sampleSelection::selection
-    if (as.list(model$call)[[1]] == "selection") {
-        model[["estimate"]] <- coefs
-        # sampleSelection::heckit
-    } else if (as.list(model$call)[[1]] == "heckit") {
+    if ("coefficients" %in% names(model)) {
         model[["coefficients"]] <- coefs
+    } else if ("estimate" %in% names(model)) {
+        model[["estimate"]] <- coefs
+    } else {
+        stop("Model does not have coefficients or estimates.")
     }
     return(model)
 }