+++ inst/tinytest/test-pkg-rstpm2.R
@@ -2,13 +2,6 @@ source("helpers.R")
 using("marginaleffects")
 requiet("rstpm2")
 
-# Basic expectation tests
-mod_simple <- rstpm2::stpm2(Surv(rectime, censrec == 1) ~ hormon, data = brcancer, df = 3)
-expect_slopes(mod_simple)
-expect_predictions(mod_simple)
-expect_hypotheses(mod_simple)
-expect_comparisons(mod_simple)
-
 ## warning: initial values uses coxph, which warns "the model contains interactions"
 suppressWarnings(test1 <- stpm2(Surv(rectime, censrec == 1) ~ hormon * x3, data = brcancer, df = 3))
 nd <- data.frame(rectime = 1000, hormon = c(0, 1), x3 = 50)